share system log
